refactoringの実施
