+++ src/main.ts
@@ -1,137 +1,68 @@
 import * as THREE from "three";
 import { GLTFLoader } from "three/addons/loaders/GLTFLoader.js";
 import { OrbitControls } from "three/addons/controls/OrbitControls.js";
-import { PerspectiveCamera } from "three";
+import {frameArea} from "./ref/frameArea.ts";
+import {resizeRendererToDisplaySize} from "./ref/resizeRender.ts";
+import {objectSetting} from "./ref/objectSetting.ts";
+
 function main() {
-  const canvas = document.querySelector("#c") as HTMLCanvasElement;
-  const renderer = new THREE.WebGLRenderer({ antialias: true, canvas });
+    const canvas = document.querySelector("#c") as HTMLCanvasElement;
+    const renderer = new THREE.WebGLRenderer({ antialias: true, canvas });
+    const fov = 45;
+    const aspect = 2; // the canvas default
+    const near = 0.1;
+    const far = 100;
+    const camera = new THREE.PerspectiveCamera(fov, aspect, near, far);
+    camera.position.set(0, 10, 20);
 
-  const fov = 45;
-  const aspect = 2; // the canvas default
-  const near = 0.1;
-  const far = 100;
-  const camera = new THREE.PerspectiveCamera(fov, aspect, near, far);
-  camera.position.set(0, 10, 20);
+    const controls = new OrbitControls(camera, canvas);
+    controls.target.set(0, 5, 0);
+    controls.update();
 
-  const controls = new OrbitControls(camera, canvas);
-  controls.target.set(0, 5, 0);
-  controls.update();
+    const scene = new THREE.Scene();
+    scene.background = new THREE.Color("black");
 
-  const scene = new THREE.Scene();
-  scene.background = new THREE.Color("black");
-
-  {
-    const planeSize = 40;
-
-    const loader = new THREE.TextureLoader();
-    const texture = loader.load(
-      "https://threejs.org/manual/examples/resources/images/checker.png",
-    );
-    texture.wrapS = THREE.RepeatWrapping;
-    texture.wrapT = THREE.RepeatWrapping;
-    texture.magFilter = THREE.NearestFilter;
-    texture.colorSpace = THREE.SRGBColorSpace;
-    const repeats = planeSize / 2;
-    texture.repeat.set(repeats, repeats);
-  }
-
-  {
-    const skyColor = 0xb1e1ff; // light blue
-    const groundColor = 0xb97a20; // brownish orange
-    const intensity = 2;
-    const light = new THREE.HemisphereLight(skyColor, groundColor, intensity);
-    scene.add(light);
-  }
-
-  {
-    const color = 0xffffff;
-    const intensity = 2.5;
-    const light = new THREE.DirectionalLight(color, intensity);
-    light.position.set(5, 10, 2);
-    scene.add(light);
-    scene.add(light.target);
-  }
-
-  function frameArea(
-    sizeToFitOnScreen: number,
-    boxSize: number,
-    boxCenter: THREE.Vector3,
-    camera: PerspectiveCamera,
-  ) {
-    const halfSizeToFitOnScreen = sizeToFitOnScreen * 0.5;
-    const halfFovY = THREE.MathUtils.degToRad(camera.fov * 0.25);
-    const distance = halfSizeToFitOnScreen / Math.tan(halfFovY);
-    // compute a unit vector that points in the direction the camera is now
-    // in the xz plane from the center of the box
-    const direction = new THREE.Vector3()
-      .subVectors(camera.position, boxCenter)
-      .multiply(new THREE.Vector3(1, 2, 1)) //カメラの角度変更
-      .normalize();
-
-    // move the camera to a position distance units way from the center
-    // in whatever direction the camera was from the center already
-    camera.position.copy(direction.multiplyScalar(distance).add(boxCenter));
-
-    // pick some near and far values for the frustum that
-    // will contain the box.
-    camera.near = boxSize / 100;
-    camera.far = boxSize * 100;
-
-    camera.updateProjectionMatrix();
-
-    // point the camera to look at the center of the box
-    camera.lookAt(boxCenter.x, boxCenter.y, boxCenter.z);
-  }
-
-  {
-    const gltfLoader = new GLTFLoader();
-    const url = "public/KC3.gltf";
-    gltfLoader.load(url, (gltf) => {
-      const root = gltf.scene;
-      scene.add(root);
-
-      // compute the box that contains all the stuff
-      // from root and below
-      const box = new THREE.Box3().setFromObject(root);
-
-      const boxSize = box.getSize(new THREE.Vector3()).length();
-      const boxCenter = box.getCenter(new THREE.Vector3());
-
-      // set the camera to frame the box
-      frameArea(boxSize * 0.5, boxSize, boxCenter, camera);
-
-      // update the Trackball controls to handle the new size
-      controls.maxDistance = boxSize * 10;
-      controls.target.copy(boxCenter);
-      controls.update();
-    });
-  }
-
-  function resizeRendererToDisplaySize(renderer: THREE.Renderer) {
-    const canvas = renderer.domElement;
-    const width = canvas.clientWidth;
-    const height = canvas.clientHeight;
-    const needResize = canvas.width !== width || canvas.height !== height;
-    if (needResize) {
-      renderer.setSize(width, height, false);
+    {
+        const skyColor = 0xb1e1ff; // light blue
+        const groundColor = 0xb97a20; // brownish orange
+        const intensity = 2;
+        const light = new THREE.HemisphereLight(skyColor, groundColor, intensity);
+        scene.add(light);
     }
 
-    return needResize;
-  }
-
-  function render() {
-    if (resizeRendererToDisplaySize(renderer)) {
-      const canvas = renderer.domElement;
-      camera.aspect = canvas.clientWidth / canvas.clientHeight;
-      camera.updateProjectionMatrix();
+    {
+        objectSetting(scene);
     }
 
-    renderer.render(scene, camera);
+
+    {
+        const gltfLoader = new GLTFLoader();
+        const url = "src/3d/model/KC3.gltf";
+        gltfLoader.load(url, (gltf) => {
+            const root = gltf.scene;
+            scene.add(root);
+            const box = new THREE.Box3().setFromObject(root);
+            const boxSize = box.getSize(new THREE.Vector3()).length();
+            const boxCenter = box.getCenter(new THREE.Vector3());
+            frameArea(boxSize * 0.5, boxSize, boxCenter, camera);
+            controls.maxDistance = boxSize * 10;
+            controls.target.copy(boxCenter);
+            controls.update();
+        });
+    }
+    function render() {
+        if (resizeRendererToDisplaySize(renderer)) {
+            const canvas = renderer.domElement;
+            camera.aspect = canvas.clientWidth / canvas.clientHeight;
+            camera.updateProjectionMatrix();
+        }
+
+        renderer.render(scene, camera);
+
+        requestAnimationFrame(render);
+    }
 
     requestAnimationFrame(render);
-  }
-
-  requestAnimationFrame(render);
 }
 
 main();
