+++ src/ref/resizeRender.ts
@@ -1,0 +1,13 @@
+import * as THREE from "three";
+
+export function resizeRendererToDisplaySize(renderer: THREE.Renderer) {
+    const canvas = renderer.domElement;
+    const width = canvas.clientWidth;
+    const height = canvas.clientHeight;
+    const needResize = canvas.width !== width || canvas.height !== height;
+    if (needResize) {
+        renderer.setSize(width, height, false);
+    }
+
+    return needResize;
+}
